experimental/HarbourMaster/project/dao/src/main/java/org/solent/com504/project/impl/dao/order/springdata/OrderRepository.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package org.solent.com504.project.impl.dao.order.springdata;

import org.solent.com504.project.impl.dao.resource.springdata.*;
import java.util.List;
import org.solent.com504.project.model.order.dto.OrderEntity;
import org.solent.com504.project.model.resource.dto.Resource;
import org.solent.com504.project.model.resource.dto.ResourceCatalog;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.query.Param;
import org.springframework.stereotype.Repository;

/**
 *
 * @author cgallen
 */
@Repository
public interface OrderRepository extends JpaRepository<OrderEntity, Long>, JpaSpecificationExecutor<OrderEntity> {

    @Query("select o from OrderEntity o  where o.uuid = :uuid")
    public List<OrderEntity> findByUuid(@Param("uuid") String uuid);

}
